+++ web/client/js/clickEvents.js
@@ -25,5 +25,18 @@ function clickEvent() {
             })
         })
     })
+
+    document.querySelectorAll(".clickEvent[data-type=remove]").forEach(el => {
+                
+        // Add them a click event:
+        el.addEventListener('click', (e) => {
+            const queryAddClass = el.getAttribute('data-classTo').split(' ').join('').split(',')
+    
+            if(!queryAddClass) return
+            queryAddClass.forEach(query => {
+                document.querySelectorAll(query).forEach(q => q.classList.remove('clicked'))
+            })
+        })
+    })
 }
 window.addEventListener('load', clickEvent)