+++ web/client/js/defaultIMG.js
@@ -1,5 +1,5 @@
 function imgDetect() {
     document.querySelectorAll('*[src]').forEach(img => {
-        img.onerror = () => img.src = '../assets/discord_logos/clyde/icon_clyde_blurple_RGB.png'
+        img.onerror = () => img.src = '../assets/discord_logos/clyde/icon_clyde_blurple_RGB.svg'
     })
 }